Moved some of the TemplateApp's MainActivity code into a abstract superclass living in the SDK
